requirements.txt
split-folders==0.5.1
keras==2.10.0
matplotlib==3.2.2
numpy==1.21.5
opencv_python==4.6.0.66
scikit_learn==1.0.2
tensorflow-gpu
